==> Code/2. Data Cleaning + EDA.ipynb <==
# Insights
- Location of the jobs
- Top hiring companies
- True proportion of data analysts vs. scientists
- Percentage of overlapped job positions
- Proportion of seniority

# Importing libraries

import re
import sys
import nltk
import unidecode
import numpy as np
import pandas as pd
import seaborn as sns
from nltk import Text
from textblob import TextBlob
from langdetect import detect
from collections import Counter
import matplotlib.pyplot as plt
from wordcloud import WordCloud
from nltk.corpus import stopwords
from googletrans import Translator
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize, RegexpTokenizer
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
%matplotlib inline
pd.set_option('display.max_colwidth', None)

df.info()

# Importing the data

df = pd.read_csv('../data/analyst_scientist_bcn_last_month.csv')

# Removing duplicate job offers (different locations only)

df.drop([159, 163, 207], inplace=True)
df.reset_index(drop=True, inplace=True)

df.head(2)

# Categorization of positions

## Proportion of job positions raw

df['title'] = df['title'].str.lower()
analysts = len(df[df['title'].str.contains("anal")])
scientists = len(df[df['title'].str.contains("scien")])
engineers = len(df[df['title'].str.contains("engine")])
print(f'Analysts: {round(analysts/df.shape[0]*100)}%')
print(f'Scientists: {round(scientists/df.shape[0]*100)}%')
print(f'Engineers: {round(engineers/df.shape[0]*100)}%')

## Overlapped job offers (cointaining analyst + scientist)

n = 0
overlapped = []
for index, job in df['title'].iteritems():
    if 'scien' in job and 'anal' in job:
        n += 1
        overlapped.append(index)
per = round(n/(analysts+scientists)*100)
print(f'{n} job positions are overlapped, {per}% of all Data Analyst & Data Scientist positions.')

df.iloc[overlapped, 0]

### Manual categorization of overlapped job position titles

Manual categorization after reading the descriptions:  
17: analyst  
37: analyst  
72: analyst  
86: scientist  
108: analyst  
150: analyst  
162: analyst  
176: analyst  

# Changing titles for categorization
df.iloc[17, 0] = 'game data analyst'
df.iloc[37, 0] = 'senior data analyst'
df.iloc[72, 0] = 'game data analyst - barcelona gaming giant'
df.iloc[86, 0] = 'data and automation scientist'
df.iloc[108, 0] = 'digital marketing data analyst'
df.iloc[150, 0] = 'senior data analyst : pricing'
df.iloc[162, 0] = 'senior data analyst : ua'
df.iloc[176, 0] = 'senior data analyst : courses'

## Categorizing job position and dropping irrelevant ones

category = []
dropped = []
for index, position in df['title'].iteritems():
    if "data scien" in position:
        category.append('data scientist')
    elif "data anal" in position:
        category.append('data analyst')
    else:
        dropped.append(index)

df.drop(dropped, inplace=True)
df.reset_index(inplace=True, drop=True)
df['position'] = category

# Exploring the data before cleaning it

## Proportion of job positions

df['position'].value_counts()

print(f"Analysts: {round(len(df[df['position'] == 'data analyst'])/df.shape[0]*100)}%")
print(f"Scientists: {round(len(df[df['position'] == 'data scientist'])/df.shape[0]*100)}%")

## Location

df['modality'].value_counts()

## Top hiring companies

df['company'].value_counts()

## Average length of descriptions

total = round(df['description'].str.split().apply(len).mean())
analysts = round(df[df['title'].str.contains("anal")]['description'].str.split().apply(len).mean())
scientists = round(df[df['title'].str.contains("scien")]['description'].str.split().apply(len).mean())
print(f'Average: {total}')
print(f'Analysts: {analysts}')
print(f'Scientists: {scientists}')

# Cleaning

## Dropping the location column

df.drop('location', axis=1, inplace=True)

## Descriptions

# Replacing line jumps with spaces
df['description'] = df['description'].str.replace('\n', ' ')
# Deleting hyphens & ¿
df['description'] = df['description'].map(lambda x: x.replace('-', '').replace('¿', ''))
# Normalising words (no accents, etc.)
df['description'] = df['description'].map(lambda x: unidecode.unidecode(x))
# Adding a space in a word if it contains a capitalised letter in between
df['description'] = df['description'].map(lambda x: re.sub(r"(?<![A-Z])(?<!^)([A-Z])",r" \1", x))
# Adding spaces before and after numbers
df['description'] = df['description'].map(lambda x: re.sub(r"([0-9]+(\.[0-9]+)?)", r" \1", x))
# Substitute n number of spaces by just one space
df['description'] = df['description'].map(lambda x: ' '.join(x.split()))
# Remove spaces before dots and after opening parenthesis
df['description'] = df['description'].map(lambda x: x.replace('( ', '(').replace(' .', '.').replace('/ ', '/'))
# Fixing some problems with the normalisation
df['description'] = df['description'].map(lambda x: x.replace('ano', 'año').replace('anos', 'años'))
# Translating the job offers in spanish
#df['description'] = df['description'].map(lambda x: str(TextBlob(x).translate(to='en'))if detect(x) == 'es' else x)

df.head(2)

# Tokenizing the descriptions + removing stopwords & other things

# Tokenising + lowercase + removing punctuation
tokenizer = RegexpTokenizer(r'\w+')
df['description'] = df['description'].map(lambda x: tokenizer.tokenize(x.lower()))
# Removing stopwords
stop_words = stopwords.words('english')
#for index, des in df['description'].iteritems():
    #df.iloc[index, 3] =  [w for w in des if not w in stop_words]
# Lemmatization
def lemma(text):
    return [WordNetLemmatizer().lemmatize(w) for w in text]
df['description'] = df['description'].apply(lemma)

df.head(2)

# Tokenizing the titles + removing stopwords & other things

# Tokenising + lowercase + removing punctuation
df['title'] = df['title'].map(lambda x: tokenizer.tokenize(x.lower()))
# Removing stopwords
for index, title in df['title'].iteritems():
    df.iloc[index, 0] = [w for w in title if not w in stop_words]

# NLTK objects for titles & descriptions

# titles
all_titles = []
for title in df['title']:
    for word in title:
        all_titles.append(word)
titles = Text(all_titles)

for index, d in df['description'].iteritems():
    d = [w.replace('visualisation', 'visualization') for w in d]
    d = [w.replace('modelling', 'modeling') for w in d]
    d = [w.replace('artificial intelligence', 'ai') for w in d]
    d = [w.replace('dashboarding', 'dashboard') for w in d]
    d = [w.replace('kpis', 'kpi') for w in d]
    d = [w.replace('report', 'reporting') for w in d]
    d = [w.replace('predict', 'prediction') for w in d]
    d = [w.replace('creative', 'creativity') for w in d]
    df['description'][index] = d

# descriptions
all_des = []
all_wrds = []
for des in df['description']:
    all_des.append(des)
    for word in des:
        all_wrds.append(word)
all_wrds = [w.replace('visualisation', 'visualization') for w in all_wrds]
all_wrds = [w.replace('modelling', 'modeling') for w in all_wrds]
all_wrds = [w.replace('artificial intelligence', 'ai') for w in all_wrds]
all_wrds = [w.replace('dashboarding', 'dashboard') for w in all_wrds]
all_wrds = [w.replace('kpis', 'kpi') for w in all_wrds]
all_wrds = [w.replace('report', 'reporting') for w in all_wrds]
all_wrds = [w.replace('predict', 'prediction') for w in all_wrds]
all_wrds = [w.replace('creative', 'creativity') for w in all_wrds]
for l in all_wrds:
    if l == 'team' and all_wrds[all_wrds.index(l) + 1] == 'working':
        all_wrds[all_wrds.index(l) + 1] = 'work'
descriptions = Text(all_wrds)

# descriptions seggregated by job position
all_des_a = []
all_des_s = []
for des in df[df['position'] == 'data analyst']['description']:
    for word in des:
        all_des_a.append(word)
for des in df[df['position'] == 'data scientist']['description']:
    for word in des:
        all_des_s.append(word)

descriptions_a = Text(all_des_a)
descriptions_s = Text(all_des_s)

df['company'].to_csv('companies.csv', index=None, header=True)

# Seniority

# Simple search in the titles
srt = 0
jrt = 0
for title in df['title']:
    if 'senior' in title or 'sr' in title:
        srt += 1
    elif 'junior' in title or 'jr' in title:
        jrt += 1

print(srt)
print(jrt)

# Simple search in the description
sr = 0
jr = 0

for des in df['description']:
    if 'senior' in des or 'sr' in des:
        sr += 1
    elif 'junior' in des:
        jr += 1

print(sr)
print(jr)

## Segmenting by seniority

seniority = []
for des in df['description']:
    if 'senior' in des or 'sr' in des:
        seniority.append('senior')
    elif 'junior' in des:
        seniority.append('junior')
    else:
        try:
            s = Text(des).concordance_list('year')[0]
            try:
                y = int(s[0][-1])
                if y < 4:
                    seniority.append('junior')
                else:
                    seniority.append('senior')
            except ValueError:
                seniority.append('unknown')
        except IndexError:
            seniority.append('unknown')

df['seniority'] = seniority
df['seniority'].value_counts()

# Keywords

## Education level

ed_level = ['bachelor', 'master', 'ph']

## Major

major = ['quantitative', 'computer science', 'engineering', 'mathematics','statistic','economics']

## Tools

tools = ['python', 'r', 'sql', 'excel', 'tableau', 'power bi', 'qlik', 'aws', 'azure', 'looker',
        'agile']

## Hard skills

hard_skills = ['programming', 'machine learning','visualization',
               'modeling','research','deep learning', 'analytical skill',
               'optimization', 'automation', 'ab', 'ai', 'dashboard',
               'kpi', 'big data', 'data mining', 'etl', 'reporting', 'segmentation',
               'business intelligence', 'natural language', 'prediction', 'algorithm']

## Soft skills

soft_skills = ['communication', 'team work', 'presentation', 'storytelling', 'decision making',
              'creativity', 'curiosity', 'leadership', 'problem solving', 'attention detail',
              'motivation', 'proactive']

## Keyword extraction

tfidf = TfidfVectorizer(ngram_range=(1,2), token_pattern='(?u)\\b\\w+\\b')
text = [' '.join(doc) for doc in df['description']]
tfidf.fit(text)
dtm = tfidf.transform(text)
dtm = pd.DataFrame(dtm.todense(), columns=tfidf.get_feature_names_out())

key_words = dtm.loc[:, dtm.columns.isin(ed_level + major + tools + hard_skills + soft_skills)]

# DF with keywords

df_key = pd.concat([df, key_words], axis=1)

df_key.drop(['description', 'title'], axis=1, inplace=True)

df_key.groupby('position').agg('mean')

df_key.to_csv('data/df_keys.csv', index=None, header=True)

test = pd.read_csv('../data/keywords.csv')
test

## To Tableau with that!

# Industries
 Extracted from Glassdoor in another notebook
