==> Code/3. Glassdoor Web Scraping.ipynb <==
import os
import re
import pandas as pd
import time
import random
import getpass
from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.chrome.service import Service
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.action_chains import ActionChains
from selenium.common.exceptions import NoSuchElementException, StaleElementReferenceException
from bs4 import BeautifulSoup
from tqdm import tqdm

df = pd.read_csv('../data/analyst.csv')

companies = df['company']

companies = companies.to_frame()

# companies = pd.read_csv('../data/companies.csv')
companies['company'] = companies['company'].str.replace('MA (Montreal Associates)', 'Montreal Associates')
emoji_pattern = re.compile("["
        u"\U0001F600-\U0001F64F"  # emoticons
        u"\U0001F300-\U0001F5FF"  # symbols & pictographs
        u"\U0001F680-\U0001F6FF"  # transport & map symbols
        u"\U0001F1E0-\U0001F1FF"  # flags (iOS)
                           "]+", flags=re.UNICODE)
companies['company'] = companies['company'].apply(str)
for index, text in companies['company'].iteritems():
    companies['company'][index] = emoji_pattern.sub(r'', text)

mail = 'praisedbethelord69@gmail.com'
pw = 'Patatacaliente_3'

def launchBrowser():
    s = Service(ChromeDriverManager().install())
    options = webdriver.ChromeOptions()
    options.add_argument('--ignore-certificate-errors')
    options.add_argument('--incognito')
    #options.add_argument("headless")
    #options.add_argument(f'user-agent={user_agent}')
    options.add_argument("start-maximized")
    options.add_experimental_option("detach", True)
    options.binary_location = '/Applications/Brave Browser.app/Contents/MacOS/Brave Browser'
    driver = webdriver.Chrome(service=s, options=options)
    driver.get("https://www.glassdoor.co.uk/index.htm")
    time.sleep(2)
    driver.find_element(By.XPATH,'//*[@id="SiteNav"]/nav/div[2]/div/div/div/button').click()
    time.sleep(2)
    driver.find_element(By.XPATH, "//*[@id='modalUserEmail']").send_keys(mail)
    time.sleep(2)
    driver.find_element(By.XPATH, "//*[@id='modalUserPassword']").send_keys(pw)
    time.sleep(2)
    driver.find_element(By.XPATH, "//*[@id='LoginModal']/div/div/div[2]/div[2]/div[2]/div/div/form/div[3]/button").click()
    return driver

def get_industry():
    location = 'Barcelona'
    driver = launchBrowser()
    time.sleep(random.randint(1,3))
    #driver.find_element(By.XPATH, "//*[@id='SiteNav']/nav[2]/div/div/div[2]/div[2]/div[1]/div/a").click()
    driver.find_element(By.XPATH, "//*[@id='SiteNav']/nav[2]/div/div/div[2]/div[2]/div[1]/div/a/span/div/h3").click()
    time.sleep(random.randint(1,3))
    for company in tqdm(companies['company']):
        time.sleep(random.randint(1,3))
        try:
            c = driver.find_element(By.XPATH, "//*[@id='sc.keyword']").get_attribute("value")
            if not c:
                driver.find_element(By.XPATH, "//*[@id='sc.keyword']").send_keys(company)
            else:
                driver.find_element(By.XPATH, "//*[@id='sc.keyword']").clear()
                driver.find_element(By.XPATH, "//*[@id='sc.keyword']").send_keys(company)

            driver.find_element(By.XPATH, "//*[@id='scBar']/div/button").click()
            time.sleep(random.randint(1,3))
            try:
                driver.find_element(By.XPATH, "//*[@id='MainCol']/div/div[1]/div/div[1]/div/div[2]/h2/a").click()
                time.sleep(random.randint(1,3))
                industries.append((company, driver.find_element(By.XPATH, "//*[@id='EIOverviewContainer']/div/div[1]/ul/li[6]/a").text))
            except NoSuchElementException:
                try:
                    industries.append((company, driver.find_element(By.XPATH, "//*[@id='EIOverviewContainer']/div/div[1]/ul/li[6]/a").text))
                except NoSuchElementException:
                    industries.append((company, 'unknown'))
        except StaleElementReferenceException:
            driver.save_screenshot(f"{random.randint(0, 100)}.png")

        

industries = []
tqdm(get_industry())

len(industries)

fix = []
i = dict(set(industries))
for c in companies['company']:
    if c == 'King':
        fix.append('Video Games')
    elif c not in i:
        fix.append('unknown')
    else:
        fix.append(i[c])
    

len(fix)

df['industry'] = fix

df.to_csv('../data/data_positions_03_05.csv', index=None, header=True)
